Last Changes

+++ part4/bloglist backend/controllers/blogs.js
@@ -4,10 +4,6 @@ const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const { userExtractor } = require('../utils/middleware')
-//
-// Lembrar de tirar o morgan das dependencias no final da parte
-// se ele realmente não for mais usado
-//
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
@@ -55,6 +51,7 @@ blogsRouter.post('/', userExtractor, async (request, response) => {
 
 blogsRouter.delete('/:id', userExtractor, async (request, response) => {
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token invalid' })
   }
@@ -62,8 +59,11 @@ blogsRouter.delete('/:id', userExtractor, async (request, response) => {
   const user = request.user
   const blog = await Blog.findById(request.params.id)
 
+  if (!blog) {
+    return response.status(204).end()
+  }
   if (!user || user._id.toString() !== blog.user.toString()) {
-    return response.status(400).json({ error: 'user ID missing or not valid' })
+    return response.status(403).json({ error: 'user ID missing or not valid' })
   }
 
   await Blog.findByIdAndDelete(request.params.id)
